src/requirements.txt
diffusers
transformers>=4.35.0
torch
accelerate
pillow
gradio>=4.0.0